Create a dump of the whole EPPO database via REST API

dataDump creates a dump of the whole EPPO database via REST API for one or more EPPO code(s).

Usage

dataDump(codesToScan, token = Sys.getenv("EPPO_token"))

Arguments

codesToScan

a character vector of one or more EPPO code(s).

token

character. Manually add your unique token, or set it inside .Renviron or using Sys.setenv(EPPO_token = "your-token").

Value

A list of flattened dataframes is returned, each dataframe containing the contents of a request converted from JSON. Each output dataframe will contain basic information, all names, taxonomy, categorization, hosts, pests and kingdom data about the input EPPO code(s), respectively. The queriedEppocode, queriedOn and queriedUrl columns in the dataframe report the input EPPO code(s), the date when the query was performed, and the queried URL, respectively.

Examples

# Create a dump of all data about Bemisia tabaci: basic information, all names, taxonomy, 
# categorization, hosts, pests and kingdom data.
dataDump("BEMITA")
#> <simpleError in doTryCatch(return(expr), name, parentenv, handler): >
#> $general
#> # A tibble: 8 × 11
#>   pestCode codeid eppocode status c_date          m_date prefname attached_infos
#>   <chr>     <int> <chr>    <chr>  <chr>           <chr>  <chr>    <named list>  
#> 1 BEMITA     5935 BEMITA   A      2002-10-28 00:… 2002-… Bemisia… <int [1]>     
#> 2 BEMITA     5935 BEMITA   A      2002-10-28 00:… 2002-… Bemisia… <int [1]>     
#> 3 BEMITA     5935 BEMITA   A      2002-10-28 00:… 2002-… Bemisia… <int [1]>     
#> 4 BEMITA     5935 BEMITA   A      2002-10-28 00:… 2002-… Bemisia… <int [1]>     
#> 5 BEMITA     5935 BEMITA   A      2002-10-28 00:… 2002-… Bemisia… <int [1]>     
#> 6 BEMITA     5935 BEMITA   A      2002-10-28 00:… 2002-… Bemisia… <int [1]>     
#> 7 BEMITA     5935 BEMITA   A      2002-10-28 00:… 2002-… Bemisia… <int [1]>     
#> 8 BEMITA     5935 BEMITA   A      2002-10-28 00:… 2002-… Bemisia… <int [1]>     
#> # ℹ 3 more variables: queriedEppocode <chr>, queriedOn <date>,
#> #   queriedUrl <glue>
#> 
#> $names
#> # A tibble: 39 × 10
#>    pestCode nameid isolang isocountry fullname  preferred author queriedEppocode
#>    <chr>     <int> <chr>   <chr>      <chr>         <int> <chr>  <chr>          
#>  1 BEMITA     7224 la      NA         Bemisia …         1 (Genn… BEMITA         
#>  2 BEMITA   211279 de      NA         Baumwoll…         0 NA     BEMITA         
#>  3 BEMITA   211280 de      NA         Tabakmot…         0 NA     BEMITA         
#>  4 BEMITA   280710 de      NA         weiße Fl…         0 NA     BEMITA         
#>  5 BEMITA     7220 en      NA         cotton w…         0 NA     BEMITA         
#>  6 BEMITA     7221 en      NA         sweet-po…         0 NA     BEMITA         
#>  7 BEMITA     7222 en      NA         tobacco …         0 NA     BEMITA         
#>  8 BEMITA   180181 en      NA         cassava …         0 NA     BEMITA         
#>  9 BEMITA   273619 en      NA         silverle…         0 NA     BEMITA         
#> 10 BEMITA     7241 es      NA         mosquita…         0 NA     BEMITA         
#> # ℹ 29 more rows
#> # ℹ 2 more variables: queriedOn <date>, queriedUrl <glue>
#> 
#> $taxonomy
#> # A tibble: 9 × 8
#>   pestCode codeid eppocode prefname  level queriedEppocode queriedOn  queriedUrl
#>   <chr>     <int> <chr>    <chr>     <int> <chr>           <date>     <glue>    
#> 1 BEMITA    57021 1ANIMK   Animalia      1 BEMITA          2025-01-15 https://d…
#> 2 BEMITA    57200 1ARTHP   Arthropo…     2 BEMITA          2025-01-15 https://d…
#> 3 BEMITA    77963 1HEXAQ   Hexapoda      3 BEMITA          2025-01-15 https://d…
#> 4 BEMITA    59061 1INSEC   Insecta       4 BEMITA          2025-01-15 https://d…
#> 5 BEMITA    58830 1HEMIO   Hemiptera     5 BEMITA          2025-01-15 https://d…
#> 6 BEMITA    73260 1STERR   Sternorr…     6 BEMITA          2025-01-15 https://d…
#> 7 BEMITA    56905 1ALEYF   Aleyrodi…     7 BEMITA          2025-01-15 https://d…
#> 8 BEMITA    57321 1BEMIG   Bemisia       8 BEMITA          2025-01-15 https://d…
#> 9 BEMITA     5935 BEMITA   Bemisia …     9 BEMITA          2025-01-15 https://d…
#> 
#> $categorization
#> # A tibble: 20 × 12
#>    pestCode nomcontinent isocode country qlist qlistlabel yr_add yr_del yr_trans
#>    <chr>    <chr>        <chr>   <chr>   <chr> <chr>       <int> <lgl>  <lgl>   
#>  1 BEMITA   America      CL      Chile   1     A1 list      2019 NA     NA      
#>  2 BEMITA   Asia         BH      Bahrain 2     A2 list      2003 NA     NA      
#>  3 BEMITA   Asia         KZ      Kazakh… 1     A1 list      2017 NA     NA      
#>  4 BEMITA   Europe       AZ      Azerba… 1     A1 list      2007 NA     NA      
#>  5 BEMITA   Europe       BY      Belarus X     Quarantin…   1994 NA     NA      
#>  6 BEMITA   Europe       GE      Georgia 1     A1 list      2018 NA     NA      
#>  7 BEMITA   Europe       MD      Moldova X     Quarantin…   2017 NA     NA      
#>  8 BEMITA   Europe       NO      Norway  X     Quarantin…   2012 NA     NA      
#>  9 BEMITA   Europe       RU      Russia  2     A2 list      2014 NA     NA      
#> 10 BEMITA   Europe       RS      Serbia  1     A1 list      2015 NA     NA      
#> 11 BEMITA   Europe       CH      Switze… 1     A1 list      2019 NA     NA      
#> 12 BEMITA   Europe       TR      Türkiye 2     A2 list      2016 NA     NA      
#> 13 BEMITA   Europe       UA      Ukraine 1     A1 list      2019 NA     NA      
#> 14 BEMITA   Europe       GB      United… 1     A1 list      2020 NA     NA      
#> 15 BEMITA   Oceania      NZ      New Ze… X     Quarantin…   2000 NA     NA      
#> 16 BEMITA   RPPO/EU      9M      EAEU    2     A2 list      2016 NA     NA      
#> 17 BEMITA   RPPO/EU      9A      EPPO    2     A2 list      1989 NA     NA      
#> 18 BEMITA   RPPO/EU      9L      EU      P     PZ Quaran…   2019 NA     NA      
#> 19 BEMITA   RPPO/EU      9L      EU      N     A1 Quaran…   2019 NA     NA      
#> 20 BEMITA   RPPO/EU      9H      OIRSA   2     A2 list      1992 NA     NA      
#> # ℹ 3 more variables: queriedEppocode <chr>, queriedOn <date>,
#> #   queriedUrl <glue>
#> 
#> $hosts
#> # A tibble: 50 × 9
#>    pestCode codeid eppocode idclass labelclass full_name         queriedEppocode
#>    <chr>     <int> <chr>      <int> <chr>      <chr>             <chr>          
#>  1 BEMITA    31422 EPHPU          1 Major host Euphorbia pulche… BEMITA         
#>  2 BEMITA    32978 GEBJA          1 Major host Gerbera jamesonii BEMITA         
#>  3 BEMITA    33334 GOSHI          1 Major host Gossypium hirsut… BEMITA         
#>  4 BEMITA    35046 IPOBA          1 Major host Ipomoea batatas   BEMITA         
#>  5 BEMITA    37151 MANES          1 Major host Manihot esculenta BEMITA         
#>  6 BEMITA    38512 NIOTA          1 Major host Nicotiana tabacum BEMITA         
#>  7 BEMITA    36921 LYPES          1 Major host Solanum lycopers… BEMITA         
#>  8 BEMITA    23377 AMABL          6 Wild/Weed  Amaranthus blito… BEMITA         
#>  9 BEMITA    23414 AMARE          6 Wild/Weed  Amaranthus retro… BEMITA         
#> 10 BEMITA    24583 ATXSE          6 Wild/Weed  Atriplex semibac… BEMITA         
#> # ℹ 40 more rows
#> # ℹ 2 more variables: queriedOn <date>, queriedUrl <glue>
#> 
#> $pests
#> # A tibble: 1 × 4
#>   pestCode queriedEppocode queriedOn  queriedUrl                                
#>   <chr>    <chr>           <date>     <glue>                                    
#> 1 BEMITA   BEMITA          2025-01-15 https://data.eppo.int/api/rest/1.0/taxon/…
#> 
#> $kingdom
#> # A tibble: 1 × 7
#>   pestCode codeid eppocode status queriedEppocode queriedOn  queriedUrl         
#>   <chr>     <int> <chr>    <chr>  <chr>           <date>     <glue>             
#> 1 BEMITA    57021 1ANIMK   A      BEMITA          2025-01-15 https://data.eppo.…
#> 

# Create a dump of all data about Aphis pomi and Leucoptera malifoliella: basic information, 
# all names, taxonomy, categorization, hosts, pests and kingdom data.
dataDump(c("APHIPO", "LEUCSC"))
#> <simpleError in doTryCatch(return(expr), name, parentenv, handler): >
#> <simpleError in doTryCatch(return(expr), name, parentenv, handler): >
#> <simpleError in doTryCatch(return(expr), name, parentenv, handler): >
#> $general
#> # A tibble: 16 × 11
#>    pestCode codeid eppocode status c_date         m_date prefname attached_infos
#>    <chr>     <int> <chr>    <chr>  <chr>          <chr>  <chr>    <named list>  
#>  1 APHIPO     5050 APHIPO   A      2002-09-17 00… 2002-… Aphis p… <int [1]>     
#>  2 APHIPO     5050 APHIPO   A      2002-09-17 00… 2002-… Aphis p… <int [1]>     
#>  3 APHIPO     5050 APHIPO   A      2002-09-17 00… 2002-… Aphis p… <int [1]>     
#>  4 APHIPO     5050 APHIPO   A      2002-09-17 00… 2002-… Aphis p… <int [1]>     
#>  5 APHIPO     5050 APHIPO   A      2002-09-17 00… 2002-… Aphis p… <int [1]>     
#>  6 APHIPO     5050 APHIPO   A      2002-09-17 00… 2002-… Aphis p… <int [1]>     
#>  7 APHIPO     5050 APHIPO   A      2002-09-17 00… 2002-… Aphis p… <int [1]>     
#>  8 APHIPO     5050 APHIPO   A      2002-09-17 00… 2002-… Aphis p… <int [1]>     
#>  9 LEUCSC    13110 LEUCSC   A      2000-12-25 00… 2000-… Leucopt… <int [1]>     
#> 10 LEUCSC    13110 LEUCSC   A      2000-12-25 00… 2000-… Leucopt… <int [1]>     
#> 11 LEUCSC    13110 LEUCSC   A      2000-12-25 00… 2000-… Leucopt… <int [1]>     
#> 12 LEUCSC    13110 LEUCSC   A      2000-12-25 00… 2000-… Leucopt… <int [1]>     
#> 13 LEUCSC    13110 LEUCSC   A      2000-12-25 00… 2000-… Leucopt… <int [1]>     
#> 14 LEUCSC    13110 LEUCSC   A      2000-12-25 00… 2000-… Leucopt… <int [1]>     
#> 15 LEUCSC    13110 LEUCSC   A      2000-12-25 00… 2000-… Leucopt… <int [1]>     
#> 16 LEUCSC    13110 LEUCSC   A      2000-12-25 00… 2000-… Leucopt… <int [1]>     
#> # ℹ 3 more variables: queriedEppocode <chr>, queriedOn <date>,
#> #   queriedUrl <glue>
#> 
#> $names
#> # A tibble: 49 × 10
#>    pestCode nameid isolang isocountry fullname  preferred author queriedEppocode
#>    <chr>     <int> <chr>   <lgl>      <chr>         <int> <chr>  <chr>          
#>  1 APHIPO     4708 la      NA         Aphis po…         1 DeGeer APHIPO         
#>  2 APHIPO     4700 da      NA         grøn æbl…         0 NA     APHIPO         
#>  3 APHIPO   213257 de      NA         grüne Ap…         0 NA     APHIPO         
#>  4 APHIPO     4702 en      NA         green ap…         0 NA     APHIPO         
#>  5 APHIPO     4703 en      NA         permanen…         0 NA     APHIPO         
#>  6 APHIPO     4718 es      NA         afido de…         0 NA     APHIPO         
#>  7 APHIPO     4719 es      NA         pulgón d…         0 NA     APHIPO         
#>  8 APHIPO     4720 es      NA         pulgón v…         0 NA     APHIPO         
#>  9 APHIPO     4724 fa      NA         schatte …         0 NA     APHIPO         
#> 10 APHIPO     4722 fi      NA         omenakir…         0 NA     APHIPO         
#> # ℹ 39 more rows
#> # ℹ 2 more variables: queriedOn <date>, queriedUrl <glue>
#> 
#> $taxonomy
#> # A tibble: 17 × 8
#>    pestCode codeid eppocode prefname level queriedEppocode queriedOn  queriedUrl
#>    <chr>     <int> <chr>    <chr>    <int> <chr>           <date>     <glue>    
#>  1 APHIPO    57021 1ANIMK   Animalia     1 APHIPO          2025-01-15 https://d…
#>  2 APHIPO    57200 1ARTHP   Arthrop…     2 APHIPO          2025-01-15 https://d…
#>  3 APHIPO    77963 1HEXAQ   Hexapoda     3 APHIPO          2025-01-15 https://d…
#>  4 APHIPO    59061 1INSEC   Insecta      4 APHIPO          2025-01-15 https://d…
#>  5 APHIPO    58830 1HEMIO   Hemipte…     5 APHIPO          2025-01-15 https://d…
#>  6 APHIPO    73260 1STERR   Sternor…     6 APHIPO          2025-01-15 https://d…
#>  7 APHIPO    57100 1APHIF   Aphidid…     7 APHIPO          2025-01-15 https://d…
#>  8 APHIPO    57101 1APHIG   Aphis        8 APHIPO          2025-01-15 https://d…
#>  9 APHIPO     5050 APHIPO   Aphis p…     9 APHIPO          2025-01-15 https://d…
#> 10 LEUCSC    57021 1ANIMK   Animalia     1 LEUCSC          2025-01-15 https://d…
#> 11 LEUCSC    57200 1ARTHP   Arthrop…     2 LEUCSC          2025-01-15 https://d…
#> 12 LEUCSC    77963 1HEXAQ   Hexapoda     3 LEUCSC          2025-01-15 https://d…
#> 13 LEUCSC    59061 1INSEC   Insecta      4 LEUCSC          2025-01-15 https://d…
#> 14 LEUCSC    59211 1LEPIO   Lepidop…     5 LEUCSC          2025-01-15 https://d…
#> 15 LEUCSC    59367 1LYONF   Lyoneti…     6 LEUCSC          2025-01-15 https://d…
#> 16 LEUCSC    59235 1LEUCG   Leucopt…     7 LEUCSC          2025-01-15 https://d…
#> 17 LEUCSC    13110 LEUCSC   Leucopt…     8 LEUCSC          2025-01-15 https://d…
#> 
#> $categorization
#> # A tibble: 9 × 12
#>   pestCode nomcontinent isocode country  qlist qlistlabel yr_add yr_del yr_trans
#>   <chr>    <chr>        <chr>   <chr>    <chr> <chr>       <int> <lgl>  <lgl>   
#> 1 APHIPO   Africa       EG      Egypt    1     A1 list      2018 NA     NA      
#> 2 APHIPO   America      CL      Chile    1     A1 list      2019 NA     NA      
#> 3 LEUCSC   Africa       EG      Egypt    1     A1 list      2018 NA     NA      
#> 4 LEUCSC   America      AR      Argenti… 1     A1 list      2019 NA     NA      
#> 5 LEUCSC   America      BR      Brazil   1     A1 list      2018 NA     NA      
#> 6 LEUCSC   America      CA      Canada   X     Quarantin…   2019 NA     NA      
#> 7 LEUCSC   America      CL      Chile    1     A1 list      2019 NA     NA      
#> 8 LEUCSC   America      US      United … X     Quarantin…   1995 NA     NA      
#> 9 LEUCSC   RPPO/EU      9D      COSAVE   1     A1 list      2021 NA     NA      
#> # ℹ 3 more variables: queriedEppocode <chr>, queriedOn <date>,
#> #   queriedUrl <glue>
#> 
#> $hosts
#> # A tibble: 8 × 9
#>   pestCode queriedEppocode queriedOn  queriedUrl         codeid eppocode idclass
#>   <chr>    <chr>           <date>     <glue>              <int> <chr>      <int>
#> 1 APHIPO   APHIPO          2025-01-15 https://data.eppo…     NA NA            NA
#> 2 LEUCSC   LEUCSC          2025-01-15 https://data.eppo…  37021 MABSD          1
#> 3 LEUCSC   LEUCSC          2025-01-15 https://data.eppo…  51515 1BETG          9
#> 4 LEUCSC   LEUCSC          2025-01-15 https://data.eppo…  52076 1CSCG          9
#> 5 LEUCSC   LEUCSC          2025-01-15 https://data.eppo…  29214 CYDOB          9
#> 6 LEUCSC   LEUCSC          2025-01-15 https://data.eppo…  41523 PRNAV          9
#> 7 LEUCSC   LEUCSC          2025-01-15 https://data.eppo…  54453 1PYUG          9
#> 8 LEUCSC   LEUCSC          2025-01-15 https://data.eppo…  42266 PYUCO          9
#> # ℹ 2 more variables: labelclass <chr>, full_name <chr>
#> 
#> $pests
#> # A tibble: 2 × 4
#>   pestCode queriedEppocode queriedOn  queriedUrl                                
#>   <chr>    <chr>           <date>     <glue>                                    
#> 1 APHIPO   APHIPO          2025-01-15 https://data.eppo.int/api/rest/1.0/taxon/…
#> 2 LEUCSC   LEUCSC          2025-01-15 https://data.eppo.int/api/rest/1.0/taxon/…
#> 
#> $kingdom
#> # A tibble: 2 × 7
#>   pestCode codeid eppocode status queriedEppocode queriedOn  queriedUrl         
#>   <chr>     <int> <chr>    <chr>  <chr>           <date>     <glue>             
#> 1 APHIPO    57021 1ANIMK   A      APHIPO          2025-01-15 https://data.eppo.…
#> 2 LEUCSC    57021 1ANIMK   A      LEUCSC          2025-01-15 https://data.eppo.…
#>