Query all services, a single service or multiple services in the EPPO database via REST API (version that never fails)

queryTheWholeEPPOSafe queries the EPPO database via REST API by specifying the name(s) of the required service(s), to retrieve basic information, all names, taxonomy, categorization, hosts, pests and kingdom data about an EPPO code. queryTheWholeEPPOSafe is a modified version of queryTheWholeEPPO that never fails.

Usage

queryTheWholeEPPOSafe(...)

Arguments

...

a function that takes the same arguments as .f. By default, .f = queryTheWholeEPPO.

The safely(.f) function from purrr/tidyverse package in R wraps a function to capture errors and side effects, i.e. creates a modified version of .f that never fails.

Value

A function that takes the same arguments as .f, but returns a different value: a list with components result and error. If the function succeeds, result contains the returned value and error is NULL. If an error occurred, error is an error object and result is either NULL or otherwise. If the function succeeds, result contains a list of flattened dataframes, each dataframe containing the contents of a request converted from JSON. Based on the query, the output dataframe will contain basic information, all names, taxonomy, categorization, hosts, pests or kingdom data about the input EPPO code. The queriedEppocode, queriedOn and queriedUrl columns in the dataframe report the input EPPO code, the date when the query was performed, and the queried URL, respectively.

Examples

# Get all information about Bemisia tabaci: basic information, all names, taxonomy, 
# categorization, hosts pests, and kingdom data.
queryTheWholeEPPOSafe("BEMITA")
#> <simpleError in doTryCatch(return(expr), name, parentenv, handler): >
#> $result
#> $result$general
#> # A tibble: 8 × 10
#>   codeid eppocode status c_date   m_date prefname attached_infos queriedEppocode
#>    <int> <chr>    <chr>  <chr>    <chr>  <chr>    <named list>   <chr>          
#> 1   5935 BEMITA   A      2002-10… 2002-… Bemisia… <int [1]>      BEMITA         
#> 2   5935 BEMITA   A      2002-10… 2002-… Bemisia… <int [1]>      BEMITA         
#> 3   5935 BEMITA   A      2002-10… 2002-… Bemisia… <int [1]>      BEMITA         
#> 4   5935 BEMITA   A      2002-10… 2002-… Bemisia… <int [1]>      BEMITA         
#> 5   5935 BEMITA   A      2002-10… 2002-… Bemisia… <int [1]>      BEMITA         
#> 6   5935 BEMITA   A      2002-10… 2002-… Bemisia… <int [1]>      BEMITA         
#> 7   5935 BEMITA   A      2002-10… 2002-… Bemisia… <int [1]>      BEMITA         
#> 8   5935 BEMITA   A      2002-10… 2002-… Bemisia… <int [1]>      BEMITA         
#> # ℹ 2 more variables: queriedOn <date>, queriedUrl <glue>
#> 
#> $result$names
#> # A tibble: 39 × 9
#>    nameid isolang isocountry fullname           preferred author queriedEppocode
#>     <int> <chr>   <chr>      <chr>                  <int> <chr>  <chr>          
#>  1   7224 la      NA         Bemisia tabaci             1 (Genn… BEMITA         
#>  2 211279 de      NA         Baumwollmottensch…         0 NA     BEMITA         
#>  3 211280 de      NA         Tabakmottenschild…         0 NA     BEMITA         
#>  4 280710 de      NA         weiße Fliege               0 NA     BEMITA         
#>  5   7220 en      NA         cotton whitefly            0 NA     BEMITA         
#>  6   7221 en      NA         sweet-potato whit…         0 NA     BEMITA         
#>  7   7222 en      NA         tobacco whitefly           0 NA     BEMITA         
#>  8 180181 en      NA         cassava whitefly           0 NA     BEMITA         
#>  9 273619 en      NA         silverleaf whitef…         0 NA     BEMITA         
#> 10   7241 es      NA         mosquita blanca d…         0 NA     BEMITA         
#> # ℹ 29 more rows
#> # ℹ 2 more variables: queriedOn <date>, queriedUrl <glue>
#> 
#> $result$taxonomy
#> # A tibble: 9 × 7
#>   codeid eppocode prefname       level queriedEppocode queriedOn  queriedUrl    
#>    <int> <chr>    <chr>          <int> <chr>           <date>     <glue>        
#> 1  57021 1ANIMK   Animalia           1 BEMITA          2025-01-15 https://data.…
#> 2  57200 1ARTHP   Arthropoda         2 BEMITA          2025-01-15 https://data.…
#> 3  77963 1HEXAQ   Hexapoda           3 BEMITA          2025-01-15 https://data.…
#> 4  59061 1INSEC   Insecta            4 BEMITA          2025-01-15 https://data.…
#> 5  58830 1HEMIO   Hemiptera          5 BEMITA          2025-01-15 https://data.…
#> 6  73260 1STERR   Sternorrhyncha     6 BEMITA          2025-01-15 https://data.…
#> 7  56905 1ALEYF   Aleyrodidae        7 BEMITA          2025-01-15 https://data.…
#> 8  57321 1BEMIG   Bemisia            8 BEMITA          2025-01-15 https://data.…
#> 9   5935 BEMITA   Bemisia tabaci     9 BEMITA          2025-01-15 https://data.…
#> 
#> $result$categorization
#> # A tibble: 20 × 11
#>    nomcontinent isocode country        qlist qlistlabel   yr_add yr_del yr_trans
#>    <chr>        <chr>   <chr>          <chr> <chr>         <int> <lgl>  <lgl>   
#>  1 America      CL      Chile          1     A1 list        2019 NA     NA      
#>  2 Asia         BH      Bahrain        2     A2 list        2003 NA     NA      
#>  3 Asia         KZ      Kazakhstan     1     A1 list        2017 NA     NA      
#>  4 Europe       AZ      Azerbaijan     1     A1 list        2007 NA     NA      
#>  5 Europe       BY      Belarus        X     Quarantine …   1994 NA     NA      
#>  6 Europe       GE      Georgia        1     A1 list        2018 NA     NA      
#>  7 Europe       MD      Moldova        X     Quarantine …   2017 NA     NA      
#>  8 Europe       NO      Norway         X     Quarantine …   2012 NA     NA      
#>  9 Europe       RU      Russia         2     A2 list        2014 NA     NA      
#> 10 Europe       RS      Serbia         1     A1 list        2015 NA     NA      
#> 11 Europe       CH      Switzerland    1     A1 list        2019 NA     NA      
#> 12 Europe       TR      Türkiye        2     A2 list        2016 NA     NA      
#> 13 Europe       UA      Ukraine        1     A1 list        2019 NA     NA      
#> 14 Europe       GB      United Kingdom 1     A1 list        2020 NA     NA      
#> 15 Oceania      NZ      New Zealand    X     Quarantine …   2000 NA     NA      
#> 16 RPPO/EU      9M      EAEU           2     A2 list        2016 NA     NA      
#> 17 RPPO/EU      9A      EPPO           2     A2 list        1989 NA     NA      
#> 18 RPPO/EU      9L      EU             P     PZ Quaranti…   2019 NA     NA      
#> 19 RPPO/EU      9L      EU             N     A1 Quaranti…   2019 NA     NA      
#> 20 RPPO/EU      9H      OIRSA          2     A2 list        1992 NA     NA      
#> # ℹ 3 more variables: queriedEppocode <chr>, queriedOn <date>,
#> #   queriedUrl <glue>
#> 
#> $result$hosts
#> # A tibble: 50 × 8
#>    codeid eppocode idclass labelclass full_name       queriedEppocode queriedOn 
#>     <int> <chr>      <int> <chr>      <chr>           <chr>           <date>    
#>  1  31422 EPHPU          1 Major host Euphorbia pulc… BEMITA          2025-01-15
#>  2  32978 GEBJA          1 Major host Gerbera jameso… BEMITA          2025-01-15
#>  3  33334 GOSHI          1 Major host Gossypium hirs… BEMITA          2025-01-15
#>  4  35046 IPOBA          1 Major host Ipomoea batatas BEMITA          2025-01-15
#>  5  37151 MANES          1 Major host Manihot escule… BEMITA          2025-01-15
#>  6  38512 NIOTA          1 Major host Nicotiana taba… BEMITA          2025-01-15
#>  7  36921 LYPES          1 Major host Solanum lycope… BEMITA          2025-01-15
#>  8  23377 AMABL          6 Wild/Weed  Amaranthus bli… BEMITA          2025-01-15
#>  9  23414 AMARE          6 Wild/Weed  Amaranthus ret… BEMITA          2025-01-15
#> 10  24583 ATXSE          6 Wild/Weed  Atriplex semib… BEMITA          2025-01-15
#> # ℹ 40 more rows
#> # ℹ 1 more variable: queriedUrl <glue>
#> 
#> $result$pests
#> # A tibble: 1 × 3
#>   queriedEppocode queriedOn  queriedUrl                                         
#>   <chr>           <date>     <glue>                                             
#> 1 BEMITA          2025-01-15 https://data.eppo.int/api/rest/1.0/taxon/BEMITA/pe…
#> 
#> $result$kingdom
#> # A tibble: 1 × 6
#>   codeid eppocode status queriedEppocode queriedOn  queriedUrl                  
#>    <int> <chr>    <chr>  <chr>           <date>     <glue>                      
#> 1  57021 1ANIMK   A      BEMITA          2025-01-15 https://data.eppo.int/api/r…
#> 
#> 
#> $error
#> NULL
#>