Merge taxonomy and kingdom data into a single dataframe

taxonomyRanked merges taxonomy and kingdom information about the queried EPPO code into a single dataframe.

Usage

taxonomyRanked(taxonomy, kingdom)

Arguments

taxonomy

a dataframe containing taxonomy data about the queried EPPO code.

kingdom

a dataframe containing kingdom data about the queried EPPO code.

Value

A dataframe containing taxonomy and kingdom information about the queried EPPO code.

Examples

# Get taxonomy and kingdom data about Bemisia tabaci:
taxonomy <- queryTheEPPOForService("BEMITA", service = "taxonomy")
kingdom <- queryTheEPPOForService("BEMITA", service = "kingdom")

# Merge taxonomy and kingdom information about Bemisia tabaci into a single dataframe:
taxonomyRanked(taxonomy, kingdom)
#> # A tibble: 9 × 8
#>   rank    codeid eppocode prefname   level queriedEppocode queriedOn  queriedUrl
#>   <chr>    <int> <chr>    <chr>      <int> <chr>           <date>     <glue>    
#> 1 kingdom  57021 1ANIMK   Animalia       1 BEMITA          2025-01-15 https://d…
#> 2 NA       57200 1ARTHP   Arthropoda     2 BEMITA          2025-01-15 https://d…
#> 3 NA       77963 1HEXAQ   Hexapoda       3 BEMITA          2025-01-15 https://d…
#> 4 NA       59061 1INSEC   Insecta        4 BEMITA          2025-01-15 https://d…
#> 5 NA       58830 1HEMIO   Hemiptera      5 BEMITA          2025-01-15 https://d…
#> 6 NA       73260 1STERR   Sternorrh…     6 BEMITA          2025-01-15 https://d…
#> 7 NA       56905 1ALEYF   Aleyrodid…     7 BEMITA          2025-01-15 https://d…
#> 8 NA       57321 1BEMIG   Bemisia        8 BEMITA          2025-01-15 https://d…
#> 9 NA        5935 BEMITA   Bemisia t…     9 BEMITA          2025-01-15 https://d…